extract `CEditorImage` into separate header
