+++ src/game/editor/editor.cpp
@@ -31,6 +31,8 @@
 #include <game/generated/client_data.h>
 #include <game/localization.h>
 
+#include <game/editor/mapitems/image.h>
+
 #include "auto_map.h"
 #include "editor.h"
