+++ src/game/editor/mapitems/map_io.cpp
@@ -11,6 +11,8 @@
 #include <game/gamecore.h>
 #include <game/mapitems_ex.h>
 
+#include "image.h"
+
 template<typename T>
 static int MakeVersion(int i, const T &v)
 {
