+++ src/game/editor/popups.cpp
@@ -12,6 +12,7 @@
 #include <limits>
 
 #include <game/client/ui_scrollregion.h>
+#include <game/editor/mapitems/image.h>
 
 #include "editor.h"
